backend/requirements.txt
django==2.2.8
djangorestframework
mysqlclient
Pillow
requests
xlwt
xlrd
redis
django-redis
jieba
celery
pycryptodome
uwsgi
